% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzAna.R
\name{anaDedup}
\alias{anaDedup}
\title{remove the duplicates of file analysis by AI.}
\usage{
anaDedup(dt, rpcols = c("file_id", "anatype"))
}
\arguments{
\item{dt}{data of analysis, data.table}

\item{rpcols}{columns which have repated lines.}
}
\description{
which data would be retained depends on time of analysis and state of analysis.
if analysised already! retain the first datetime of analysised.
if analysising or not analysis yet! retain the last datetime of initiate analysis.
}
