% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkBase.R
\name{api.jkdic}
\alias{api.jkdic}
\title{获取Teable下jkdic的api地址}
\usage{
api.jkdic(rtype = "debt", take = 500, skip = 0)
}
\arguments{
\item{rtype}{财报类型，包含debt,benefit,cash,main，用于search}

\item{take}{api返回数据量，默认500行数据}

\item{skip}{指定api跳过的数量，用于分页获取数据}
}
\value{
返回可使用的GET类数据
}
\description{
获取Teable下jkdic的api地址
}
