% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkBase.R
\name{dbc.local}
\alias{dbc.local}
\title{建立数据库本地数据库通道}
\usage{
dbc.local(dbname = "finace", dbtype = "postgres")
}
\arguments{
\item{dbname}{数据库名称,默认finace}

\item{dbtype}{数据库类型，目前支持postgres 与 mysql}
}
\value{
返回已建立的数据库通道
}
\description{
建立数据库本地数据库通道
}
