% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAbsr}
\alias{getAbsr}
\title{get absolute address of a file/folder.}
\usage{
getAbsr(adr)
}
\arguments{
\item{adr}{address of file/folder.}
}
\value{
file address
}
\description{
addr which begins with *, is icloud address in mac/win.
addr which begins with space or ~,is full address of mac or linux.
addr which begins with space or ~,is full address of mac or linux.
}
\details{
additional property:credbile identifies file/folder exist or not
}
