% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCloudr}
\alias{getCloudr}
\title{get address of icloud.}
\usage{
getCloudr()
}
\value{
NULL OR address of icloud folder.
}
\description{
icloud address depends OS system: windows or mac
}
\details{
address in windows:C:/Users/Administrator/iCloudDrive
address in macos:/Users/longsa/Library/Mobile Documents/com~apple~CloudDocs
}
