% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HailContextManager-class.R
\name{HailContext}
\alias{HailContext}
\alias{hail_context}
\alias{hail}
\title{Create a HailContext}
\usage{
hail_context()

hail()
}
\description{
To use Hail features it is necessary to create a connection, the function 
\code{hail_context()} create this connection and returns a jobj of 
class \code{is.Hail.HailContext}. The \code{hail()} funciton is just a wrapper of 
\code{hail_context()}.
}
\examples{
\dontrun{
hail()
}
}
