% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-utils.R
\name{install_hail}
\alias{install_hail}
\title{Install Hail}
\usage{
install_hail(version)
}
\arguments{
\item{version}{A string with the Hail version}
}
\description{
Install versions of Hail for use with local HailContext.
}
