% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkBase.R
\name{jkCheck}
\alias{jkCheck}
\title{校验jqka类型}
\usage{
jkCheck(x = NULL, jktype = NULL)
}
\arguments{
\item{x}{待校验的值}

\item{jktype}{jqka校验类型}
}
\value{
F or T
}
\description{
支持的类型: rtype, period and drc.
}
\details{
rtype类型包含的值: debt,benefit,cash,main
period类型包含的值: year, report, simple
drc类型包含的值: w2l, l2w
}
