% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkGet.R
\name{jkConvert}
\alias{jkConvert}
\title{转换同花顺财报数据，宽数据<->长数据}
\usage{
jkConvert(dt, drc = "w2l")
}
\arguments{
\item{dt}{data.table格式数据}

\item{drc}{转换方向, wide to long(w2l), long to wide(l2w)}
}
\value{
返回转换后的数据
}
\description{
转换同花顺财报数据，宽数据<->长数据
}
