% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkGet.R
\name{jkCorl}
\alias{jkCorl}
\title{链接mysql获取jkdic表数据，扩展jqka数据得到自定义字段数据}
\usage{
jkCorl(dt, rtype = "debt", corlby = "subject")
}
\arguments{
\item{dt}{同花顺jqka数据表data.table}

\item{rtype}{同花顺数据类型}

\item{corlby}{链接字段，需要为dt与jkdic共同名称的字段}
}
\value{
返回关联了jkdic后的数据data.table
}
\description{
jkdic数据来源:默认Teable，若失败则读本地postgres
}
