% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkGet.R
\name{jkMillis}
\alias{jkMillis}
\title{将指定列转换为数值，并转换为百万单位}
\usage{
jkMillis(dt, rtype = "debt")
}
\arguments{
\item{dt}{data.table格式数据}

\item{rtype}{jqka财报数据类型：debt,benefit,cash,main}
}
\value{
返回转换后的数据
}
\description{
若包含百分号%，则不转换。
}
