% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkGet.R
\name{jkRead}
\alias{jkRead}
\title{读取同花顺财报文件、简单清洗并关联数据库中的dics}
\usage{
jkRead(jkdir = ".", rtype = "debt")
}
\arguments{
\item{jkdir}{为文件地址(包含文件名)}

\item{rtype}{jqka财报数据类型:debt,benefit,cash,main}
}
\value{
返回读取并简单清理后的财报数据data.table
}
\description{
读取同花顺财报文件、简单清洗并关联数据库中的dics
}
