% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkGet.R
\name{jksRep}
\alias{jksRep}
\title{下载、读取并清洗计算同花顺jqka财报数据}
\usage{
jksRep(
  symbs,
  rtp = "debt",
  prd = "year",
  pwd = "./",
  dw = 0,
  gr = 0.15,
  xr = 0.08
)
}
\arguments{
\item{symbs}{上市公司代码}

\item{rtp}{jqka财报类型：debt,benefit,cash,main}

\item{prd}{jqka财报日期：year(年报), report(报告期汇总), simple(单报告期)}

\item{pwd}{jqka财报文件保存地址}

\item{dw}{是否重新下载jqka财报文件}

\item{gr}{为公司收益增长率，默认15%}

\item{xr}{为10年期国债收益率,默认8%}
}
\value{
返回财报数据列表list，包括vertical、horizontal两个表
}
\description{
下载、读取并清洗计算同花顺jqka财报数据
}
