% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkGet.R
\name{jkUrl}
\alias{jkUrl}
\title{获取同花顺jqka财报的链接地址}
\usage{
jkUrl(symb, rtype = "debt", period = "year")
}
\arguments{
\item{symb}{单个股票代码}

\item{rtype}{财报类型，包括debt/benefit/cash/main}

\item{period}{财报期数，包括year/report/simple三种}
}
\value{
返回同花顺财报文件的链接地址
}
\description{
获取同花顺jqka财报的链接地址
}
