% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkGet.R
\name{jkYoy}
\alias{jkYoy}
\title{计算指定列wcols的年数值变动率}
\usage{
jkYoy(dt, wcols = "name", dcol = "date", vcol = "value")
}
\arguments{
\item{dt}{data.table数据}

\item{wcols}{forget the usage of this argument}

\item{dcol}{date column}

\item{vcol}{value column}
}
\value{
返回计算后的数据表
}
\description{
计算指定列wcols的年数值变动率
}
