% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send}
\alias{send}
\title{Sending
 
Utlitlies for sending data somewhere and still being able to use it.}
\usage{
send(x, dest)
}
\arguments{
\item{x}{}

\item{dest}{}
}
\description{
Sending
 
Utlitlies for sending data somewhere and still being able to use it.
}
