% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkGet.R
\name{jkCalc}
\alias{jkCalc}
\title{计算各财报中需要增加的额外数据及估值}
\usage{
jkCalc(dt, rtype = "debt", grate = 1, xrate = 0.08)
}
\arguments{
\item{dt}{data.table格式数据}

\item{rtype}{jqka财报类型：debt,benefit,cash,main}

\item{grate}{为公司收益增长率，默认10%（增长五年），后续按3%增长}

\item{xrate}{为10年期国债收益率,默认8%}
}
\value{
返回处理后的数据
}
\description{
计算各财报中需要增加的额外数据及估值
}
