% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkBase.R
\name{jkCheck}
\alias{jkCheck}
\title{check out jqka's data type.}
\usage{
jkCheck(x = NULL, jktype = NULL)
}
\arguments{
\item{x}{item need be checking.}

\item{jktype}{kind of jqka's type: rtype/period/drc}
}
\value{
F or T
}
\description{
check type: rtype, period and drc.
rtype: debt,benefit,cash,main
period: year, report, simple
drc: w2l, l2w
}
